Add Filtering (#2)
Filtering support is now available, although it is still an experimental feature.

Snapshot testing module had a couple of changes, so now it produces constant unique snapshot file names. Previously the snapshot name was based on the test name, but now it is independent. It was very annoying, because if you rename the test you also need to rename the snapshot file.

Parser module is renamed to Extract.